modules/apigateway/lambda_authorizer/requirements.txt
PyJWT==2.7.0
requests==2.31.0
cryptography>=41.0.7
pytest>=6.0.0,<8.0.0
